backend/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
pdf2image
langchain
langchain-google-genai
chromadb
reportlab
resend
pydantic-settings
python-dotenv
google-generativeai
tiktoken
pymupdf
python-docx
pillow
fastapi-limiter
sentence-transformers
pypdf
sqlmodel
cryptography
fpdf2
slowapi
groq
email-validator
